+++ Samples/tvOS-Swift/tvOS-Swift/AppDelegate.swift
@@ -12,8 +12,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         SentrySDK.start { options in
             options.dsn = "https://a92d50327ac74b8b9aa4ea80eccfb267@o447951.ingest.sentry.io/5428557"
             options.debug = true
-            options.logLevel = SentryLogLevel.verbose
-            options.attachStacktrace = true
             options.sessionTrackingIntervalMillis = 5_000
         }